CutSolution class gone - functionality folded into CutCircuit
